backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.5
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
httpx==0.27.2
azure-storage-blob==12.23.1
python-pptx==1.0.2
langchain==0.3.3
langchain-core==0.3.10
langchain-community==0.3.2
langchain-ollama==0.2.0
langchain-postgres==0.0.12
langgraph==0.2.34
langsmith==0.1.131
ollama==0.3.3
